chore: move livesplit typings in its own types file
