+++ components/types/livesplit.ts
@@ -1,0 +1,12 @@
+/**
+ * LiveSplit-related types
+ */
+
+import { Seconds } from "./types"
+
+export interface LiveSplitTimeCalcEntry {
+    contractId: string
+    time: Seconds
+    location: string
+    isCompleted: boolean
+}
